requirements.txt
streamlit
langchain
langchain-google-genai
beautifulsoup4
requests
python-dotenv
chromadb
